+++ src/classifier/IFeatureClassifier.hpp
@@ -13,8 +13,8 @@ public:
 
 	//***** Classifier *****
 	virtual bool train(const std::vector<std::vector<Eigen::RowVectorXd>>& datasets) = 0;
-	virtual bool classify(const Eigen::RowVectorXd& sample, uint32_t& classid) = 0;
-	virtual bool classify(const Eigen::RowVectorXd& sample, uint32_t& classid, std::vector<double>& distance, std::vector<double>& probability) = 0;
+	virtual bool classify(const Eigen::RowVectorXd& sample, size_t& classid) = 0;
+	virtual bool classify(const Eigen::RowVectorXd& sample, size_t& classid, std::vector<double>& distance, std::vector<double>& probability) = 0;
 
 	//***** XML *****
